Create a 16x16 grid of square divs
Create 16 containers (row) which each contains a container (col) with 16 identical square divs. The squares all maintain the aspect ratio of a square when shrinking and (not yet) expands if there is extra room on screen. The body covers the whole viewport and there is a minimum size to the grid.