Create grid with biggest square div in viewport
Change style.css to mostly center grid on page. Add padding and background-color as decoration. Switch names of row and col in script.js to reflect correct naming convention (row is horizontal and col is vertical). Change to allow grid to grow but stop when it exceeds viewport by limiting width of a specific col (size of 1 square). The exact value which limits the width is derived from the smaller dimension of viewport (i.e. calculate using the length of viewport if it is smaller than width of viewport). The flexbox property allows grid to shrink but not grow as it sets maximum size upon loading initial viewport